clean create-vite
